# openapi/glass-game-panel-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Glass Game Panel Control Plane API
  version: 0.1.5-stub
  summary: GlassGamePanel control plane (draft)
  description: |
    Draft control-plane API for GlassGamePanel (`0.1.5-stub`). Humans and agents
    share this API. A UI is another client. Wings is the current executor; a
    later driver may sit behind the same verbs. This draft is not a production deploy.

    **Defaults:**
    - Documented production base: `https://panel-api.glasshosting.com/v1` (not published with this draft)
    - `serverId` is a Glass UUID. The Wings UUID is `wings_server_id`
    - Auth is Portal OIDC. Portal mints agent tokens. Agent tokens cannot mint child tokens
    - `agents:manage` is for people, not for agent tokens
    - Console is a ticket plus WebSocket. No Authorization header on the WebSocket handshake
    - Agents use the files API. Optional SFTP for people is later
    - RCON is deferred to v1.1
    - `network` is reserved. There is no allocate or release path in this draft
    - Rate limits: 60 mutate requests and 300 read requests per minute per token
    - Repository: `FuzzySpace/GlassGamePanel`

    **Migration:**
    - GlassHosting-owned Pterodactyl first. A panel the customer operates is later
    - Default mode is `inventory_only`
    - Rollback window default is 72 hours (design stub)
    - Servers the caller does not own stay excluded. `include_managed_inventory` is forced false in this build

    **Refuse:**
...
    not handed to agents.

...

...

...
servers:
...
paths:
  /servers:
    get:
      tags: [Servers]
      operationId: listServers
      summary: List servers visible to the caller
      security:
        - BearerAuth: [servers:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ServerStatus'
      responses:
        '200':
          description: Cursor page of servers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerPage'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Servers]
      operationId: createServer
      summary: Create a server (job completes with an IPv4 and IPv6 claim; refused servers return 403 managed_inventory_denied; default executor does not dial Wings)
      security:
        - BearerAuth: [servers:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '202':
          description: Accepted; provision job started
          headers:
            Location:
              description: Job URL
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Servers]
      operationId: getServer
      summary: Get one server
      security:
        - BearerAuth: [servers:read]
      responses:
        '200':
          description: Server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Servers]
      operationId: updateServer
      summary: Patch server metadata or limits (not power; refused servers return 403 managed_inventory_denied)
      security:
        - BearerAuth: [servers:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
      responses:
        '200':
          description: Updated server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Servers]
      operationId: deleteServer
      summary: Delete server (async; refused servers return 403 managed_inventory_denied)
      security:
        - BearerAuth: [servers:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Delete job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}/power:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    post:
      tags: [Power]
      operationId: serverPower
      summary: Start, stop, restart, or kill (refused servers return 403 managed_inventory_denied)
      security:
        - BearerAuth: [power]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PowerAction'
      responses:
        '202':
          description: Power transition accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '409':
          description: Invalid transition for current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}/console:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Console]
      operationId: serverConsoleWs
      summary: Console stream (WebSocket upgrade)
      description: |
        Ticket plus WebSocket only. Clients mint
        `POST /servers/{id}/console-ticket` and connect with the ticket as a
        query parameter on the WebSocket URL. Authorization Bearer on the
        handshake is not supported and must be rejected.

        Ticket TTL is at most 120 seconds (`max_ttl_seconds`), single-use, and
        bound to `serverId` plus the minting subject. Refused servers are
        rejected on mint and on upgrade.

        Not a JSON HTTP body. Frames are text console I/O. Shell escape is
        refused. A command allowlist is later.
      # WebSocket auth is the ticket query parameter only (no BearerAuth)
      security: []
      responses:
        '101':
          description: Switching Protocols (WebSocket)
        '400':
          description: Non-upgrade HTTP request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}/console-ticket:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    post:
      tags: [Console]
      operationId: createConsoleTicket
      summary: Mint a short-lived console WebSocket ticket (refused servers return 403 managed_inventory_denied)
      description: |
        Returns a single-use ticket bound to this `serverId` and the minting
        subject. `expires_at` must be at most now plus `max_ttl_seconds` (120).
        Refused servers return 403 `managed_inventory_denied`. WebSocket clients
        authenticate with the ticket only (no Bearer on upgrade).
      security:
        - BearerAuth: [console]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsoleTicket'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}/files:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Files]
      operationId: listFiles
      summary: List a directory under the server jail (refused servers return 403 managed_inventory_denied)
      description: |
        Agents use the files API only. Optional SFTP credentials for people
        are outside the v1 agent path.
      security:
        - BearerAuth: [files]
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            default: /
            description: Absolute path within server root jail
      responses:
        '200':
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileList'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}/files/content:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Files]
      operationId: readFile
      summary: Read file bytes (size-capped)
      security:
        - BearerAuth: [files]
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [Files]
      operationId: writeFile
      summary: Write an allowlisted jar or zip under mods/ or plugins/
      description: |
        Refused servers return 403 `managed_inventory_denied` with Wings dispatch 0.
        `PANEL_WINGS_EXECUTOR=noop`, an empty `PANEL_WINGS_ALLOWLIST`, and a UUID
        that is not on that list return 403 and do not dial. When the executor is
        `real`, the daemon origin and token are set, and the path UUID is
        allowlisted, this dials Wings `POST /api/servers/{uuid}/files/write`.
        The path must stay under `mods/` or `plugins/` and end in `.jar` or `.zip`.
        `../`, host paths, and unit paths are refused before any dial.
        The body must be a zip local-file archive. Evidence is `X-Request-Id`,
        `X-Glass-Wings-HTTP-Status`, and `wings_dialed` / `wings_http_status`.
      security:
        - BearerAuth: [files]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: path
          in: query
          required: true
          schema:
            type: string
            description: Volume path under mods/ or plugins/ (.jar or .zip)
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Wings write completed (evidence JSON; Wings status is wings_http_status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMutateResult'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Files]
      operationId: deleteFile
      summary: Delete one allowlisted jar or zip under mods/ or plugins/
      description: |
        Same refuse rule, allowlist, executor, and path policy as writeFile.
        An allowlisted delete dials Wings `POST /api/servers/{uuid}/files/delete`
        for that one file. It does not delete host paths.
      security:
        - BearerAuth: [files]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wings delete completed (evidence JSON)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMutateResult'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}/backups:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Backups]
      operationId: listBackups
      summary: List backups for server
      security:
        - BearerAuth: [backups]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Backup page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupPage'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Backups]
      operationId: createBackup
      summary: Start backup job (progress via Jobs / SSE)
      security:
        - BearerAuth: [backups]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupCreate'
      responses:
        '202':
          description: Backup job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        default:
          $ref: '#/components/responses/Error'
  /servers/{serverId}/backups/{backupId}/restore:
    parameters:
      - $ref: '#/components/parameters/ServerId'
      - name: backupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Backups]
      operationId: restoreBackup
      summary: Restore a backup (async; refused servers return 403 managed_inventory_denied)
      security:
        - BearerAuth: [backups]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Restore job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        default:
          $ref: '#/components/responses/Error'
...
  /agent-tokens:
    get:
      tags: [AgentTokens]
      operationId: listAgentTokens
      summary: List agent tokens (metadata only; secret never re-shown)
      security:
        - BearerAuth: [agents:manage]
      responses:
        '200':
          description: Token metadata list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentToken'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [AgentTokens]
      operationId: createAgentToken
      summary: Mint scoped agent token (TTL required)
      security:
        - BearerAuth: [agents:manage]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTokenCreate'
      responses:
        '201':
          description: Token created; secret returned once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTokenCreated'
        default:
          $ref: '#/components/responses/Error'
...
  /migrations/ptero:
    post:
      tags: [Migrations]
      operationId: startPteroMigration
      summary: Start Ptero panel import (easy migration)
      description: |
        Imports servers from a source Pterodactyl panel into GlassGamePanel.

        Servers the caller does not own are skipped. `include_managed_inventory` is
        forced false in this build. A client boolean, `managed_inventory_allowlist`,
        and `founder_yes_attestation_id` together are still rejected (no opt-in
        via request fields). Default is exclude.

        Flow: validate source, discover servers, nodes, and eggs, build a remap
        plan, copy files and backups when configured, then move pointers.
        An optional rollback window follows.

        GlassHosting-owned Pterodactyl only in v1. A panel the customer operates
        is later. Calls should use `mode=inventory_only` or `mode=dry_run`
        (schema default `inventory_only`). Rollback window default is 72 hours
        (design stub).

        `GLASSPANEL_MIGRATE_MODE_MAX` is `inventory_only` or `dry_run` only.
        `import` and `import_and_cutover` are refused. Source read uses
        `PTERO_SOURCE_API_URL` and `PTERO_SOURCE_API_TOKEN` (names only; the
        request token is not stored or forwarded). The Wings executor stays `noop`.

        The source Application API token must be scoped to read servers and
        allocations. Write scope on the source is not required for a plan.
      security:
        - BearerAuth: [migrations:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PteroMigrationCreate'
      responses:
        '202':
          description: Migration job accepted
          headers:
            Location:
              description: Migration resource URL
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migration'
        default:
          $ref: '#/components/responses/Error'
  /migrations:
    get:
      tags: [Migrations]
      operationId: listMigrations
      summary: List migration jobs visible to the caller
      security:
        - BearerAuth: [migrations:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/MigrationStatus'
      responses:
        '200':
          description: Migration page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPage'
        default:
          $ref: '#/components/responses/Error'
  /migrations/{migrationId}:
    parameters:
      - $ref: '#/components/parameters/MigrationId'
    get:
      tags: [Migrations]
      operationId: getMigration
      summary: Get migration job status + remap summary
      security:
        - BearerAuth: [migrations:read]
      responses:
        '200':
          description: Migration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migration'
        default:
          $ref: '#/components/responses/Error'
...
  /migrations/{migrationId}/confirm:
    parameters:
      - $ref: '#/components/parameters/MigrationId'
    post:
      tags: [Migrations]
      operationId: confirmMigration
      summary: Confirm remap and start copy/cutover
      security:
        - BearerAuth: [migrations:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Execute phase started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migration'
        default:
          $ref: '#/components/responses/Error'
...
  /jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Jobs]
      operationId: getJob
      summary: Get async job status
      security:
        - BearerAuth: [servers:read]
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT from GlassPortal OIDC. Scopes are space-delimited.
        Portal mints agent tokens (`POST /agent-tokens` is a staff or owner
        session). Agents cannot mint child tokens. Do not use long-lived
        administrator keys. Mutual TLS between Portal and Wings is transport
        and is outside this OpenAPI file.
...